pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.0",
    "psycopg2-binary>=2.9.11",
]
